Move Video.get_frame_by_index to MediaType
